requirements.txt
flask
flask-cors
mysql-connector-python
razorpay
qrcode
python-dotenv